Add animWalk on enmies and gauge cards

+++ Assets/Scripts/PlayerShoot/Projectile/Projectile.cs
@@ -36,7 +36,7 @@ public class Projectile : MonoBehaviour
     {
         return usedCard.cooldown;
     }
-    public int GetAmmo()
+    public float GetAmmo()
     {
         return usedCard.ammo;
     }
